pyproject.toml
[project]
name = "pixel-hawk"
version = "0.1.0"
description = "Pixel Hawk - WPlace paint project change tracker"
authors = [
    { name = "Liz Balbuena" }
]
license = "MIT"
requires-python = ">=3.14"
dependencies = [
    "loguru>=0.7.3",
    "pillow>=12.0.0",
    "httpx>=0.28.0",
    "ruamel-yaml>=0.19.1",
    "tortoise-orm>=0.25.0",
]

[project.scripts]
pixel-hawk = "pixel_hawk.main:main"

[tool.ruff]
line-length = 120

[build-system]
requires = ["uv_build"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=7.0.0",
    "ty>=0.0.17",
    "aerich[toml]>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
addopts = "--cov=pixel_hawk --cov-report=term-missing --cov-fail-under=95 -q"
python_files = ["tests/*.py"]
testpaths = ["tests"]

[tool.coverage.run]
data_file = ".pytest_cache/.coverage"

[tool.ty]
rules = { all = "error" }
src = { exclude = ["tests"] }
environment = { root = ["./src"] }

[tool.aerich]
tortoise_orm = "pixel_hawk.db.TORTOISE_ORM"
location = "./src/pixel_hawk/migrations"
src_folder = "./."
